Refactor Ecoflow Client and Database Interaction
- Removed unused SVG logo file from the project.
- Updated GetDeviceAllParameters method in EcoflowClient to be asynchronous and improved response parsing.
- Modified EcoflowPostgreDbContext to accept a connection string and ensure proper configuration.
- Refactored DataHelper to use dependency injection for EcoflowPostgreDbContext, improving database interaction.
- Updated DbHelper methods to handle decimal values for battery SOC and voltage instead of integers.
- Adjusted DeviceDataFetcher to utilize the refactored DataHelper for fetching device data asynchronously.
- Changed BatteryDataTracker properties from int to decimal for better precision in battery data representation.

+++ EcoflowShared/Models/BatteryDataTracker.cs
@@ -4,8 +4,8 @@
         public string devicename { get; set; }
         public string serialnumber { get; set; }
         public DateTime datetime { get; set; }
-        public int soc { get; set; }
-        public int extrabattery1soc { get; set; }
-        public int extrabattery2soc { get; set; }
+        public decimal soc { get; set; }
+        public decimal extrabattery1soc { get; set; }
+        public decimal extrabattery2soc { get; set; }
         public int voltage { get; set; }
     }
